Conflicts:
	qemu/exec-all.h

+++ qemu/translate-all.c
@@ -155,7 +155,7 @@ int cpu_gen_code(CPUArchState *env, TranslationBlock *tb, int *gen_code_size_ptr
 /* The cpu state corresponding to 'searched_pc' is restored.
  */
 int cpu_restore_state(TranslationBlock *tb,
-                      CPUArchState *env, unsigned long searched_pc)
+                      CPUArchState *env, uintptr_t searched_pc)
 {
 /**
  *  The retranslation mechanism interferes with S2E's instrumentation.
@@ -166,7 +166,7 @@ int cpu_restore_state(TranslationBlock *tb,
 
     TCGContext *s = &tcg_ctx;
     int j;
-    unsigned long tc_ptr;
+    uintptr_t tc_ptr;
 #ifdef CONFIG_PROFILER
     int64_t ti;
 #endif
@@ -193,7 +193,7 @@ int cpu_restore_state(TranslationBlock *tb,
 #endif
 
     /* find opc index corresponding to search_pc */
-    tc_ptr = (unsigned long)tb->tc_ptr;
+    tc_ptr = (uintptr_t)tb->tc_ptr;
     if (searched_pc < tc_ptr)
         return -1;
